Slate aside plugin normalizing
