+++ src/components/SlateEditor/plugins/aside/index.tsx
@@ -6,20 +6,14 @@
  *
  */
 
-import {
-  afterTextBlockElement,
-  firstTextBlockElement,
-  lastTextBlockElement,
-  textBlockElements,
-} from '../../utils/normalizationHelpers';
 import React from 'react';
-import { Descendant, Editor, Element, Node, Path, Text, Transforms } from 'slate';
-import { jsx } from 'slate-hyperscript';
+import { Descendant, Editor, Element } from 'slate';
 import { RenderElementProps } from 'slate-react';
+import { jsx } from 'slate-hyperscript';
+import { defaultTextBlockNormalizer } from '../../utils/normalizationHelpers';
 import { SlateSerializer } from '../../interfaces';
 import SlateAside from './SlateAside';
 import { getAsideType } from './utils';
-import { TYPE_PARAGRAPH } from '../paragraph';
 
 export const TYPE_ASIDE = 'aside';
 
@@ -60,57 +54,11 @@ export const asidePlugin = (editor: Editor) => {
   };
 
   editor.normalizeNode = entry => {
-    const [node, path] = entry;
+    const [node] = entry;
 
     if (Element.isElement(node) && node.type === TYPE_ASIDE) {
-      if (node.children.length === 0) {
-        Transforms.insertNodes(editor, jsx('element', { type: TYPE_PARAGRAPH }, [{ text: '' }]), {
-          at: [...path, 0],
-        });
-        return;
-      }
-
-      const firstChild = node.children[0];
-      if (Element.isElement(firstChild)) {
-        if (!firstTextBlockElement.includes(firstChild.type)) {
-          Transforms.insertNodes(editor, jsx('element', { type: TYPE_PARAGRAPH }, [{ text: '' }]), {
-            at: [...path, 0],
-          });
-          return;
-        }
-      }
-
-      const lastChild = node.children[node.children.length - 1];
-      if (Element.isElement(lastChild)) {
-        if (!lastTextBlockElement.includes(lastChild.type)) {
-          Transforms.insertNodes(editor, jsx('element', { type: TYPE_PARAGRAPH }, [{ text: '' }]), {
-            at: [...path, node.children.length],
-          });
-          return;
-        }
-      }
-
-      for (const [child, childPath] of Node.children(editor, path)) {
-        if (Text.isText(child)) {
-          Transforms.wrapNodes(editor, jsx('element', { type: TYPE_PARAGRAPH }), { at: childPath });
-          return;
-        }
-        if (Element.isElement(child) && !textBlockElements.includes(child.type)) {
-          Transforms.unwrapNodes(editor, {
-            at: childPath,
-          });
-          return;
-        }
-      }
-
-      const next = Editor.next(editor, { at: path });
-      if (next) {
-        const [nextNode, nextPath] = next;
-        if (!Element.isElement(nextNode) || !afterTextBlockElement.includes(nextNode.type)) {
-          Transforms.wrapNodes(editor, jsx('element', { type: TYPE_PARAGRAPH }), { at: nextPath });
-          return;
-        }
-      }
+      defaultTextBlockNormalizer(editor, entry, nextNormalizeNode);
+      return;
     }
     nextNormalizeNode(entry);
   };
